Adds check for PSS option around PSS warning
Currently, the check for whether PSS reading is available is always run, even if we don't want to use it.

This change checks whether it was requested before checking.

Arguably it might also be desirable to check whether it was requested or just the default, and fail if it was specifically requested, but that's a larger change.